+++ bot/bot.py
@@ -5,7 +5,7 @@ VERSION = version_info(
 	name="captain",
 	major=3,
 	minor=1,
-	patch=0,
+	patch=1,
 	release="stable"
 )